backend/requirements/web_api.requirements.txt
fastapi==0.91.0
pydantic==1.10.4
uvicorn==0.20.0
PyJWT==2.6.0